% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEnhFit.r
\name{backEnhFit}
\alias{backEnhFit}
\title{backEnhFit}
\usage{
backEnhFit(
  tmp.df,
  df,
  initialVars,
  possibleVars,
  m.i,
  eff.ind.inside,
  max.df.spline,
  eff.inside.dates,
  trap,
  model.info,
  fits,
  plot.file,
  option,
  bsplBegDt,
  bsplEndDt
)
}
\arguments{
\item{tmp.df}{The reduced data frame originating from \code{df} containing 
only efficiency-trial dates; i.e., those with non-\code{NA}
\code{nReleased} values.}

\item{df}{The data frame for a specific \code{TrapPositionID} containing 
efficiency-trial information and covariates, if available, at the time of 
fitting enhanced efficiency trials in \code{eff_model.r} (or 
\code{F.efficiency.model }).}

\item{initialVars}{The set of possible variables available for model
inclusion, following application of the 90% presence rule.  These variables
form a subset of the \code{possibleVars}.}

\item{possibleVars}{The set of all available variables for possible inclusion
to the fitting of an enhanced efficiency model.  Usually include
efficiency-trial variables, environmental covariate variables, and
CAMP-collected variables.}

\item{m.i}{An integer containing the number of rows in data frame 
\code{tmp.df}.}

\item{eff.ind.inside}{A logical vector of length equal to the number of rows
of data frame \code{df}, expressing which \code{batchDates} fall within the
temporal data-range for which estimation occurs.}

\item{max.df.spline}{An integer expressing the highest allowable spline
degree.}

\item{eff.inside.dates}{A \code{POXIX}-type vector of length 2, housing the 
first and last \code{batchDates} for which vector \code{eff.ind.inside} is 
\code{TRUE}.}

\item{trap}{An alphanumeric \code{TrapPositionID} value, as itemized within
\code{df} (and \code{tmp.df}).}

\item{model.info}{A list containing information derived from an immediate
preceding call to function \code{plotbsSpline}.  Used to help get the
current fit iteration going.}

\item{fits}{A list object of length equal to the number of 
\code{TrapPositionID}s requiring a model fit.  Ultimately passed to the 
boostrapping routine.}

\item{plot.file}{The name of the file prefix under which output is to be 
saved.  Set to \code{NA} to plot to the plot window.}

\item{option}{A graphical option for displaying spline fits.  Typically (and hard-coded) set to \code{2}.}

\item{bsplBegDt}{The first date, via the spline-1959-1960 paradigm, to which
all efficiency years and dates collapse.}

\item{bsplEndDt}{The last date, via the spline-1959-1960 paradigm, to which
all efficiency years and dates collapse.}
}
\value{
Several objects, all housed within a list.  

\describe{
  \item{fit}{The final fit for the current \code{TrapPositionID} of interest.}  
  \item{model.info}{List of model information resulting from the application of \code{backEnhFit} to each unique \code{TrapPositionID} in \code{df}.}  
  \item{fits}{List of fits resulting from the application of \code{backEnhFit} to each unique \code{TrapPositionID} in \code{df}.}  
  \item{covarStringPlot}{A record of considered variables necessary model plotting outside of the function.  }
  \item{interimVars1Num}{Set of variables remaining for model inclusion when both CAMP \code{waterTemp_C} and Environmental-covariate {temp_C} are present.}  
  \item{interimVars2Num}{Set of variables remaining for model inclusion when both CAMP \code{discharge_cfs} and Environmental-covariate \code{flow_cfs} are present.}
}
}
\description{
For a given set of covariates, fit a genearlized additive model
  via a backwards-selection paradigm, allowing for both selection of
  confounding covariates, as well as variable temporal B-splines.
}
\details{
Function \code{bachEnhFit} is the workhorse function that fits the
  backwards-fitting algorithm for enhanced-efficiency splines.  It utilizes
  tests of deviances to identify variables for possible exclusion from an
  initial model containing all avalabile covariates, and smaller subets
  following the removal of the most non-significant covariate identified on
  the most recent preceding model fitting.
  
  An \eqn{alpha} value of 0.10 is used to test for possible stopping of the
  model;  i.e., if all included covariates in a model have p-values less than
  0.10, all covariates are retained, and the model-fitting procedure stops. 
  Note that the value of 0.10 is set within the function via initial
  parameter \code{pCutOff}.
}
\examples{
\dontrun{
out <- backEnhFit(tmp.df,
                  df,
                  initialVars,
                  possibleVars,
                  m.i,
                  eff.ind.inside,
                  max.df.spline,
                  eff.inside.dates,
                  trap,
                  model.info,
                  fits,
                  plot.file,
                  option,
                  bsplBegDt,
                  bsplEndDt)
}
}
\seealso{
\code{eff_model_enh}, \code{plotbsSpline}
}
\author{
WEST Inc.
}
