% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campR.r
\docType{data}
\name{betas}
\alias{betas}
\title{Beta coefficients resulting from the fitting of trap- (\code{trapPositionID}
or \code{subSiteID}) and river-specific (\code{site}) enhanced-efficiency
models.}
\format{
A data frame with 40 rows and 24 variables.

\describe{ A dataset containing beta estimates resulting from the fitting of 
trap-specific generalized additive logistic models enhanced efficiency models
in the Central Valley of California, USA.  All coefficients are on the logit 
scale, and interpret as the average positive increase (negative decrease) in 
log-odds of efficiency for every one-unit increase (decrease) in the measured
variable.  Positive values communicated average incraese, while negative 
values communicate average decrease.  For example, a value of +1.50 for
\code{turbidity_ntu} communicates that for every one-ntu increase in
turbidity is associated with a +1.50 increase in the log-odds of efficiency. 
The variables are as follows:

\itemize{
  \item \code{subsiteID}. Trap identifier.
  \item \code{(Intercept)}. Estimated intercept.  
  \item \code{bdMeanNightProp}. Estimated effect due to proportion of fishing at night.  
  \item \code{bdMeanMoonProp}. Estimated effect due to proportion of fishing with moon in the sky.
  \item \code{bdMeanForkLength}. Estimated effect due to mean fork length.  
  \item \code{flow_cfs}. Estimated effect due to flow, in units of cfs, as recorded in the Environmental Covariate database.
  \item \code{temp_c}. Estimated effect due to water temperature, in units of C, as recorded in the Environmental Covariate database.  
  \item \code{discharge_cfs}.Estimated effect due to flow, in units of cfs, as recorded in the CAMP \code{mdb}.
  \item \code{waterDepth_cm}.Estimated effect due to water depth, in units of cm, as recorded in the CAMP \code{mdb}.  
  \item \code{waterDept_ft}. Estimated effect due to water dept, in units of ft, as recorded in the CAMP \code{mdb}.  
  \item \code{airTemp_C}. Estimated effect due to air temperature, in units of C, as recorded in the CAMP \code{mdb}.  
  \item \code{airTemp_F}. Estimated effect due to air temperature, in units of F, as recorded in the CAMP \code{mdb}.  
  \item \code{turbidity_ntu}. Estimated effect due to turbidity, in units of ntu, as recorded in the CAMP \code{mdb}.  
  \item \code{waterVel_fts}. Estimated effect due to water velocity, in units of ft/s, as recorded in the CAMP \code{mdb}.  
  \item \code{waterTemp_C}. Estimated effect due to water temperature, in units of C, as recorded in the CAMP \code{mdb}.  
  \item \code{waterTemp_F}. Estimated effect due to water temperature, in units of F, as recorded in the CAMP \code{mdb}.  
  \item \code{lightPenetration_ntu}. Estimated effect due to light penetration, in units of ntu, as recorded in the CAMP \code{mdb}.  
  \item \code{dissolvedOxygen_mgL}. Estimated effect due to dissolved oxygen, in units of mg/L, as recorded in the CAMP \code{mdb}.  
  \item \code{conductivity_mgL}. Estimated effect due to conductivity, in units of mg/L, as recorded in the CAMP \code{mdb}.  
  \item \code{barometer_inHg}. Estimated effect due to barometer, in units of inches Hg, as recorded in the CAMP \code{mdb}.  
  \item \code{precipLevel_qual}. Estimated effect due to precipitation level, on a 4-unit quantitative scale, as recorded in the CAMP \code{mdb}.  
  \item \code{threshold}. The number of distinct efficiency trials with no missing covariates utilized to estimate the betas for the given \code{subsiteID}.
  \item \code{available}. The number of distinct efficiency trials considered for beta estimation for the given \code{subsiteID}.
  \item \code{Stage}. Ubiquitously set to value of 'Final Model Betas'.  
}
}
}
\usage{
data(betas)
}
\description{
Beta coefficients resulting from the fitting of trap- (\code{trapPositionID}
or \code{subSiteID}) and river-specific (\code{site}) enhanced-efficiency
models.
}
\keyword{datasets}
