% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAstroStats.r
\name{buildAstroStats}
\alias{buildAstroStats}
\title{buildAstroStats}
\usage{
buildAstroStats(release.visit, visit.df)
}
\arguments{
\item{release.visit}{A data frame containing release data obtained from the
release-data SQL sequence.}

\item{visit.df}{A data frame containing visits obtained from the catch-data
SQL sequence.}
}
\value{
A dataframe with one row for all unique \code{batchDates} included 
  within \code{min.date} and \code{max.date}, inclusive.  All variables, as 
  provided via \code{varVec}, have a non-\code{NA} value.
}
\description{
Given a particular trap sequence, calculate the proportion of
  time spent fishing during nighttime and while the moon was up.
}
\details{
The \code{buildAstronStats} function first queries the CAMP database
  for all relevant visits via the \code{Visits} table, as well as the 
  astronomical sun and moon rise and set times, contained within table 
  \code{Dates}.  The "Not Fishing" data are also obtained, via a query to 
  table \code{TempSumUnmarkedByTrap_Run_Final} in the underlying CAMP 
  \code{mdb}.  Because of this, the catch SQL sequence must be run prior to 
  the same of the efficiency.
  
  The function emulates the calculation of \code{SampleMinutes} performed by 
  the catch SQL sequence.  This is done in order to calculate denominators 
  for calculation of the proportion of time spent fishing at night and while 
  the moon is up.  The \code{SampleMinutes} field from 
  \code{TempSumUnmarkedByTrap_Run_Final} was not utilized because the 
  astronomical metrics need to calculate fishing with respect to the included
  fishing instances in the calculation of efficiency.  These include fishing 
  instances regardless of the value of the \code{includeCatchID} field.  The 
  value of \code{SampleMinutes} in the \code{TempSumUnmarkedByTrap_Run_Final}
  calculated for catch estimation excludes bad fishing, as identified via the
  \code{includecatchID} field.  Therefore, re-estimation of 
  \code{SampleMinutes}, based on efficiency concerns, must be performed. 
  These \code{SampleMinutes} are called \code{JasonSampleMinutes}, so as to 
  differentiate. Values of this variable equals \code{-99} for the first 
  visit of a trapping sequence, and \code{-88} in the case of a time frame 
  greater than the gap in fishing length. Currently, the fishing-gap length 
  is set via global variable \code{fishingGapMinutes}, and equals 
  \code{10080} minutes.
  
  Calculation of astronomical proportions utilized \code{campR} functions 
  \code{makeSkinnyTimes} and \code{getTimeProp}.
}
\examples{
\dontrun{
df <- buildAstroStats(release.visit=release.visit,
                      visit.df=visit.df)
}
}
\seealso{
\code{makeSkinnyTimes}, \code{getTimeProp}
}
\author{
WEST Inc.
}
