% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campR.r
\docType{package}
\name{campR}
\alias{campR}
\title{campR: R routines for the CAMP R platform.}
\section{Description}{
 Allow annual estimation of both catch and efficiency of
  juvenile Chinook salmon in the Central Valley of California.
}

