% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campR.r
\docType{data}
\name{canal}
\alias{canal}
\title{Average cubic feet per second recorded from the total dishcarge diverted to
the two side canals immediately prior to the flow of the Scaramento River
through the Red Bluff Diversion Dam.}
\format{
A data frame with 366 rows and 2 variables.

\describe{A dataset containing a row for each day of the leap-year in the format of \code{mm-dd}, and another column holding the average 
cubic feet per second through the two diversion canals on the Sacramento River immediately before the Red Bluff Diversion Dam.  Daily 
values obtained from the average of 11 years' worth of recorded data, 2002-2013.  Data obtained from Felipe Carrillo.  
The variables are as follows:

\itemize{
  \item \code{monthDay}. Month-day identifier in the form \code{mm-dd}.  Note entry for \code{2-29}.  
  \item \code{average_cfs}. Estimated diversion daily flow in cubic feet per second.  
}
}
}
\usage{
data(canal)
}
\description{
Average cubic feet per second recorded from the total dishcarge diverted to
the two side canals immediately prior to the flow of the Scaramento River
through the Red Bluff Diversion Dam.
}
\keyword{datasets}
