% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chuck_zeros.R
\name{chuck.zeros}
\alias{chuck.zeros}
\title{chuck.zeros}
\usage{
chuck.zeros(b, e, beg.buff, end.buff, df)
}
\arguments{
\item{b}{An integer representing the number of zeros and \code{NA} trapping
instances to keep before the first non-zero and non-\code{NA} catch record.
Always set to zero.}

\item{e}{An integer representing the number of zeros and \code{NA} trapping
instances to keep after the last non-zero and non-\code{NA} catch record.
Always set to zero.}

\item{beg.buff}{An integer.  The actual number of zeros and \code{NA}
  trapping instances prior to the first non-zero and non-\code{NA} catch
record in the original data frame \code{df}.}

\item{end.buff}{An integer.  The actual number of zeros and \code{NA} 
trapping instances after the last non-zero and non-\code{NA} catch record 
in the original data frame \code{df}.}

\item{df}{A data frame containing catch records for one particular trap's 
temporal sequence.}
}
\value{
A data frame, necessarily a subset of the data frame \code{df} 
  provided to the function, with all zero and \code{NA} catch records removed
  before the first, and after the last, valid catch record.
}
\description{
Removes zeros before the first non-zero catch, and the same for
  zeros following the last non-zero catch.
}
\details{
Manipulation of the \code{b} and \code{e} variables specifies the 
  number of trapping days with preceding and antecedent zeros to retain. 
  Thus, the number of zero catch records both before the first and after the 
  last caught fish is independently adjustable.
  
  Note that variable \code{n.tot} in data frame \code{df} identifies zero and
  \code{NA} catch records.
}
\examples{
\dontrun{
#   ---- Remove all zero and NA records before the first non-zero and non-NA
#   ---- catch record in record 4 in data frame catch.df.  Do the same for
#   ---- the last, which is in record 12.  
df <- chuck.zeros(0,0,4,12,catch.df)
}
}
\seealso{
\code{max.buff.days}
}
\author{
WEST Inc.
}
