% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCovar.R
\name{estCovar}
\alias{estCovar}
\title{estCovar}
\usage{
estCovar(dbCov, covName, estType, traps, obs.eff.df, xwalk, oursitevar)
}
\arguments{
\item{dbCov}{An object originating from function \code{getCAMPEnvCov}, i.e., 
preprocessed environmental data from a CAMP database.}

\item{covName}{The text string name of an environmental covariate, with its 
units appended via an underscore, e.g., \code{temperature_C}.}

\item{estType}{A single number taking on values \code{1} or \code{2}.  Number
\code{1} is used for quantitative variables, while number \code{2} is used
for quantitative.}

\item{traps}{A vector of \code{trapPositionID}s encompassing all the unique 
traps in data frame \code{obs.eff.df}.}

\item{obs.eff.df}{A data frame containing observed efficiency trials.  Note
that this data frame contains one day for all in-between days, given
efficiency trials.}

\item{xwalk}{A data frame containing lookup information tying
\code{subSiteID} to the unique identifying \code{ourSiteIDChoice1} (and
\code{ourSiteIDChoice2}) in the online postgreSQL covariate database.}

\item{oursitevar}{The integer of length one identifying the
\code{ourSiteIDChoice1} (or \code{ourSiteIDChoice2}) of the
\code{subSiteID}'s identifying station number in the postgreSQL covariate
database.}
}
\value{
Data frame \code{obs.eff.df} with the requested covariate's data
 values appended on days for which data were available.
}
\description{
Query individual covariates and their recorded units from an 
 Access CAMP database.
}
\details{
Quantitative variables are fed into a smoothing spline (via function
 \code{smooth.spline}) so as to allow for exact interpolation / prediction on
 efficiency time values.
 
 Qualitative variables, which currently only includes \code{"weather"},
 simply reports values.  No smoothing is performed.
}
\examples{
\dontrun{
estCovar(dbCov,covName,estType,traps,obs.eff.df)
}
}
\seealso{
\code{getCAMPEnvCov}
}
\author{
WEST Inc.
}
