% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCAMPEnvCov.R
\name{getCAMPEnvCov}
\alias{getCAMPEnvCov}
\title{getCAMPEnvCov}
\usage{
getCAMPEnvCov(dbCov, cov, covID, unitID)
}
\arguments{
\item{dbCov}{The data frame resulting from a query against the
\code{EnvDataRaw} table in a CAMP mdb, following unit-code clean-up via the
QryCleanEnvCov SQL sequence.}

\item{cov}{The text string name of the environmental covariate in the
\code{EnvDataRaw} table of the CAMP Access database.}

\item{covID}{The text string name of the environmental covariate units
variable tied to \code{cov} in the \code{EnvDataRaw} table of the CAMP
Access database.}

\item{unitID}{A single number describing the CAMP Access unit code appropriate
for the provided \code{cov} and \code{unitID}. These derive from lookup
table \code{luUnit} in the CAMP database.}
}
\value{
A data frame containing all the records from table \code{EnvDataRaw}
 in the CAMP database.  Only records with a \code{covID} record matching the
 \code{unitID} are returned, although all unique values encountered are saved
 as attribute \code{"uniqueUnitID"}.
}
\description{
Query individual covariates and their recorded units from an
 already queried Access CAMP database.
}
\details{
Qualitative variables, which currently only includes
 \code{"weather"}, do not subset to values with a valid \code{unitID}, if
 only because these don't exist.  Alternatively, a lookup table of all
 weather values ever recorded is used to map to one of four ordinal values of
 increasing cloudiness.
}
\examples{
\dontrun{
getCAMPEnvCov(cov,covID,unitID)
}
}
\seealso{
\code{estCovar}
}
\author{
WEST Inc.
}
