% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatchForkLength.R
\name{getCatchForkLenth}
\alias{getCatchForkLenth}
\title{getCatchForkLength - Retrieve catch data with fork length information attaches.}
\usage{
getCatchForkLenth(site, taxon, run, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{run}{The run code.  This is an integer value that corresponds to 
a run listed in table \code{luRun} which is housed inside the Access database. 
At the time of this writing, the \code{luRun} table translated run codes as:
\itemize{
  \item 1 = Spring
  \item 2 = Summer
  \item 3 = Fall
  \item 4 = Winter
  \item 5 = Late fall
  \item 6 = Mixed
}
Normally, the translation from text (e.g., "fall") to code (e.g., 3) is done 
by the camp interface.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
Either a single histogram, if variable \code{by.lifestage} is set to
  \code{FALSE}. Otherwise, a histogram for each individual life stage present
  within the data.
}
\description{
Retrieve a data frame from the CAMP Access dBAse containing
catches and fork length.  This routine includes fish from both \code{includeCatch = TRUE}
and \code{includeCatch = FALSE} catches.
}
\section{Details}{

This routine compiles fish from the CAMP Access dBase using function
  \code{F.get.indiv.fish.data}.  Fish are included from both "valid" and 
  "invalid" trapping instances where "valid" means \code{includeCatchID = TRUE}.  Thus, 
  resulting graphs can display both valid and invalid fishing instances if 
  both types of trapping instances took place between
  \code{min.date} and \code{max.date.}
  
  Observations in the output graph are jittered slightly when plotted to show 
  the number of measurements on each date.  
  
  When no valid catch records are found, a blank \code{png} is returned.
}

\section{Author}{

 Trent McDonald and Jason Mitchell
}

\examples{
\dontrun{
#   ---- Obtain graphical histograms for the American. 
site <- 57000
taxon <- 161980
run <- 3
min.date <- "2014-01-01"
max.date <- "2014-06-06"
output.file <- "American"
by.lifestage <- TRUE

F.length.frequency(site,taxon,run,min.date,max.date,output.file,by.lifestage)
}
}
\seealso{
\code{F.get.indiv.fish.data}
}
