% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_buff_days.r
\name{max_buff_days}
\alias{max_buff_days}
\title{max_buff_days}
\usage{
max_buff_days(df, trap)
}
\arguments{
\item{df}{A data frame containing catch records for one 
\code{trapPositionID}, with catch counts contained in \code{n.tot}.}

\item{trap}{A vector of length one containing the \code{trapPositionID} for 
which buffering is necessary.}
}
\value{
A vector containing three values.  The first contains the row index 
  of the first non-zero / non-\code{NA} in the catch record provided via data
  frame \code{df}. The second contains the row index of the last non-zero / 
  non-\code{NA} in the catch record provided via data frame \code{df}.  The 
  third contains the length of the vector containing 
  \code{trapVisitID}-specific fish counts (vector \code{n.tot}).
}
\description{
Identify the number of preceding and antecdedent zeros in the 
  catch record for an individual trap catch sequence.
}
\details{
Function \code{max_buff_days} is a helper function utilized in the 
  estimation of catch, and assists to identify preceding and antecedent 
  zeros.  It works in conjunction with function \code{F.chuck.zeros}.
}
\examples{
\dontrun{
#   ---- Find the first and last non-zero and non-NA catch record
#   ---- for trap "12345".
vec <- max.buff.days(df,"12345")
}
}
\seealso{
\code{F.chuck.zeros}
}
\author{
WEST Inc.
}
