% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overDphi.r
\name{overDphi}
\alias{overDphi}
\title{overDphi}
\usage{
overDphi(model, family = "binomial", type)
}
\arguments{
\item{model}{A \code{glm} object resulting from a call to function
\code{glm}.  Typically either a catch Poisson model or efficiency binomial
model.}

\item{family}{The family of exponential distributions.  Currently allows
either of \code{family="binomial"} or \code{family="poisson"}.}

\item{type}{The type of residual to calculate.  Currently allows only
\code{"pearson"}.}
}
\value{
An estimate of the overdispersion \eqn{\phi} statistic from the
  fitting of either a \code{family="binomial"} or \code{family="poisson"}
  generlized linear model.
}
\description{
For a set of covariates for which estimates have been obtained
  on non-continguous dates, expand recorded data values to fill in all dates,
  so as to create a step-like function.
}
\details{
The \code{overDphi} function excludes Pearson residuals larger in
  magnitude of 8 in the case of binomial fits.  In the case of Poisson fits,
  only the residuals falling within the 20% and 80% percentiles are retained.
  
  Estimates of the overdispersion less than one are set to one.
}
\examples{
\dontrun{
disp <- overDphi(model=fit,
                 family="binomial",
                 type="pearson")
}
}
\author{
WEST Inc.
}
