% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryEnvCovAPI.R
\name{queryEnvCovAPI}
\alias{queryEnvCovAPI}
\title{Query the Environmental Covariate API}
\usage{
queryEnvCovAPI(min.date, max.date, oursitevar, type = "D")
}
\arguments{
\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.  Queries on the Unit 
table \code{tblu} assume a start time of midnight on the day specified.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.  Queries on the Unit table \code{tblu} assume an end time 
of midnight on the day specified.}

\item{oursitevar}{An integer indicating the site identifier.  These identifiers
must be the custom ones designed specifically for the Environmental 
Covariate Database.}

\item{type}{A text string indicating if either Daily \code{"D"} or Unit 
\code{"U"} values should be queried.  At present, only "D" is implemented.}
}
\value{
A data frame of the query results, with a fields for both 
  flow and temperature values, what statistics are returned, and date or 
  time interval. The returned data frame has the following fields: 
  \itemize{
    \item \code{date} = date of the measurement in POSIXct format.
    \item \code{flow_statistic} = an integer specifying the type of statistic 
    returned for flow (1 = maximum; 2 = minimum; 3 = mean; 8 = median).
    \item \code{flow_statistic_name} = the name of the flow statistic.  
    "Maximum", "Minimum", "Mean", "Median" as specified by \code{flow_statistic}. 
    \code{flow_statistic_name} and \code{flow_statistic} are redundant, the former 
    is included for convenience. 
    \item \code{flow_cfs} = the value of flow statistic in cubic feet per second. 
    \item \code{temp_statistic} = an integer specifying the type of statistic 
    returned for flow (1 = maximum; 2 = minimum; 3 = mean; 8 = median).
    \item \code{temp_statistic_name} = the name of the flow statistic.  
    "Maximum", "Minimum", "Mean", "Median" as specified by \code{temp_statistic}. 
    \code{temp_statistic_name} and \code{temp_statistic} are redundant, the former 
    is included for convenience. 
    \item \code{temp_cfs} = the value of flow statistic in cubic feet per second. 
  }
}
\description{
Query the Environmental Covariate API for data between two
  dates.
}
\details{
For a given time interval (e.g., day or hour), a site may have more than one statistic
  recorded.  Statistics are mean, median, max, or min over the time point.  
  The statistic returned by this routine for time intervals with multiple statistics
  is determined by the following rules, applied in order of precedence: 
  \enumerate{
    \item median is selected before the mean
    \item mean is selected before the max or min
    \item if max AND min are available (but neither median or mean), 
    an average of min and max is returned
    \item if only one of max or min is available, it is returned.  
  }
}
\examples{
\dontrun{
#   ---- Query for daily results.  
tmp <- queryEnvCovAPI(min.date = "2000-01-01",
              max.date = "2010-12-31",
              oursitevar = 11,
              type = "D")
}

}
\seealso{
\code{plot2Panel}
}
\author{
Trent McDonald and Jason Mitchell
}
