% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceETrials.r
\name{reduceETrials}
\alias{reduceETrials}
\title{reduceETrials}
\usage{
reduceETrials(df, possibleVars, bsplBegDt, bsplEndDt, trap, all.ind.inside)
}
\arguments{
\item{df}{The data frame for a specific \code{TrapPositionID} containing 
efficiency-trial information and covariates, if available, at the time of 
fitting enhanced efficiency trials in \code{eff_model.r} (or 
\code{F.efficiency.model }).}

\item{possibleVars}{The set of all available variables for possible inclusion
to the fitting of an enhanced efficiency model.  Usually include
efficiency-trial variables, environmental covariate variables, and
CAMP-collected variables.}

\item{bsplBegDt}{The first date, via the spline-1959-1960 paradigm, to which
all efficiency years and dates collapse.}

\item{bsplEndDt}{The last date, via the spline-1959-1960 paradigm, to which
all efficiency years and dates collapse.}

\item{trap}{A trap for which efficiency data are available.}

\item{all.ind.inside}{A logical vector of length equal to the number of rows
of data frame \code{df}, expressing which \code{batchDates} fall within the
temporal data-range for which estimation occurs.}
}
\value{
A list containing several items.  

\describe{
  \item{df}{The original \code{df} with \code{batchDate2} appended.}
  \item{tmp.df}{The set of efficiency trials for the \code{TrapPositionID} of interest.}
  \item{m.i}{The number of efficiency trials within the time period of interest.}
  \item{initialVars}{The set of possible covariates available prior to model fitting.}
  \item{initialVarsNum}{An accounting data frame helpful in creating data frame \code{betas} following model fitting.}
  \item{all.ind.inside}{A logical vector indicating which \code{batchDates} fell within the temporal range spanned by valid efficiency trials.}
}
}
\description{
Map the set of efficiency trials to the 1959-1960 enhanced
  efficiency spline-fitting paradigm.
}
\details{
Function \code{reduceETrials} does the work of mapping several
  years' worth of efficiency trials to one common temporal period.  For
  convenience, the common temporal period is 1960.  1960 was select because
  it is a leap-year.  It is also near the \code{POSIX} temporal original date
  of \code{"1970-01-01"}.  This is helpful in rare cases in which direct
  manipulation of \code{POSIX} items is necessary, since raw numerics
  associated with \code{POSIX} variables record the number of seconds since 
  the origination date.
  
  Sometimes, the spline year paradigm for a river may stretch back into 1959.
  It never stetches into 1961.  This is important for use in function 
  \code{eff_model}, where care must be taken to ensure that the resulting 
  1960-spline fit housed in previously fit enhanced efficiency data maps 
  correctly to provided \code{min.date} and \code{max.date}, when provided
  for passage estimation.  Lack of attention here could result in re-maps to
  the wrong \code{min.date} and \code{max.date} year.
}
\examples{
\dontrun{
ans <- reduceETrials(df,possibleVars,bsplBegDt,bsplEndDt,trap,all.ind.inside)
}
}
