% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reMap.R
\name{reMap}
\alias{reMap}
\title{reMap}
\usage{
reMap(df, bd, min.date, max.date, strt.dt, end.dt)
}
\arguments{
\item{df}{Usually a \code{c0} dataframe housing the basis matrix used to fit 
an underlying enhanced efficiency model.  Thus, rows are unique 
(\code{batchDate2} in the 1959-1960 paradigm), while columns are dimensions
associated with the underlying b-spline matrix.}

\item{bd}{A character vector of length one identifying the name of the 
variable holding \code{batchDate2} to be mapped back (remapped) to the 
present.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{strt.dt}{The remapped start date associated with the current trap's 
minimum (earliest) spline date.}

\item{end.dt}{The remapped end date associated with the current trap's 
maximum (latest) spline date.}
}
\value{
A list with two dataframes.  The first, \code{c0} contains the 
  remapped \code{batchDate}s (so originating \code{batchDate2}, remapped to 
  \code{batchDate}), for the originaing b-spline basis matrix, while the 
  second, \code{allDates} also contains remapped \code{batchDate}s for all 
  days of interest.
  
  Dataframe \code{allDates} is used as the originating left-table dataframe 
  against which all other covariate dataframes are merged.
}
\description{
Map 1959-1960 efficiency-trial \code{batchDate}s back to the 
  present, as defined via \code{min.date} and \code{max.date}.
}
\examples{
\dontrun{
reMap(df,bd,min.date,max.date,strt.dt,end.dt)
}
}
\author{
WEST Inc.
}
