% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reMap2.R
\name{reMap2}
\alias{reMap2}
\title{reMap2}
\usage{
reMap2(min.date, max.date, splineDays)
}
\arguments{
\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{splineDays}{The set of spline days in the 1959-1960 paradigm on which 
enhanced efficiency models were fit.}
}
\value{
A data frame with all observed and imputed \code{efficiency} values, 
  where variable \code{gam.estimated} identifies days with imputed values.
}
\description{
Map the dates of an enhanced efficiency model fit back to the
  present, as defined via \code{min.date} and \code{max.date}.
}
\examples{
\dontrun{
reMap2( min.date,max.date,splineDays )
}
}
\seealso{
\code{reMap}
}
\author{
WEST Inc.
}
