% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableDeleter.R
\name{tableDeleter}
\alias{tableDeleter}
\title{tableDeleter}
\usage{
tableDeleter()
}
\value{
The function is silent.  It simply reports messages to the log
  file indicating deleted tables.
}
\description{
Query the CAMP.mdb for all internal
  tables, and in the case any identified "temp" tables are present, delete them, 
  so as to ensure the CAMP.mdb is as small as possible.
}
\details{
The function requires no arguments.  The location of the CAMP.mdb of
  interest is known via externally defined global variable \code{db.file}.
  
  Additionally, the current set of temp tables used for checking is
  hard-coded within the function, so this set of tables does not require
  specification.
}
\author{
WEST, Inc.
}
